+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- Gameplay: increase Shield duration to 26s to improve usefulness (tiny) — 2026-03-23
+
 ## 9.62.0
 - UI: make Auto-Pause button state clearer (tiny) — 2026-03-23